BE/django/anaconda_requirements.txt
absl-py==1.4.0
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
boto3==1.26.94
botocore==1.29.94
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
Django==4.0
django-extensions==3.2.1
djangorestframework==3.14.0
executing==1.2.0
filelock==3.9.0
fire==0.5.0
fonttools==4.39.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
huggingface-hub==0.13.2
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.2.1
mysqlclient==2.1.1
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.6
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
sahi==0.11.12
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.3
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.2.3
yolov5==7.0.9
zipp==3.15.0
